Restructure the project to implement iteration 2
To wrap the utility into distribution package with
setuptools some restructuring of the project was made.

Version 1.0.0
