+++ .gitignore
@@ -122,7 +122,7 @@ __pycache__/
 .Python
 build/
 develop-eggs/
-dist/
+#dist/
 downloads/
 eggs/
 .eggs/
@@ -149,7 +149,7 @@ pip-log.txt
 pip-delete-this-directory.txt
 
 # Unit test / coverage reports
-htmlcov/
+VitaliBekarevich/htmlcov/
 .tox/
 .nox/
 .coverage
