+++ VitaliBekarevich/rss_reader/setup.cfg
@@ -1,0 +1,14 @@
+[metadata]
+name = rss_reader
+version = 0.4.0
+
+[options]
+packages = rss_reader_p
+scripts = rss_reader.py
+install_requires =
+    requests
+    defusedxml
+
+[options.entry_points]
+console_scripts =
+    rss_reader = rss_reader:main
